utilities/insert-into-file: Extract standalone unit tests
