+++ tests/unit/models/blueprint-test.js
@@ -1598,212 +1598,6 @@ help in detail');
     });
   });
 
-  describe('insertIntoFile', function() {
-    var blueprint;
-    var ui;
-    var tmpdir;
-    var project;
-    var filename;
-
-    beforeEach(function() {
-      return mkTmpDirIn(tmproot).then(function(dir) {
-        tmpdir = dir;
-        blueprint = new Blueprint(basicBlueprint);
-        ui        = new MockUI();
-        project   = new MockProject();
-        // normally provided by `install`, but mocked here for testing
-        project.root = tmpdir;
-        blueprint.project = project;
-        filename = 'foo-bar-baz.txt';
-      });
-
-    });
-
-    afterEach(function() {
-      return remove(tmproot);
-    });
-
-    it('will create the file if not already existing', function() {
-      var toInsert = 'blahzorz blammo';
-
-      return blueprint.insertIntoFile(filename, toInsert)
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.contain(toInsert);
-          expect(result.originalContents).to.equal('', 'returned object should contain original contents');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was modified');
-          expect(contents).to.equal(result.contents, 'returned object should contain contents');
-        });
-    });
-
-    it('will insert into the file if it already exists', function() {
-      var toInsert = 'blahzorz blammo';
-      var originalContent = 'some original content\n';
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert)
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal(originalContent + toInsert, 'inserted contents should be appended to original');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was modified');
-        });
-    });
-
-    it('will not insert into the file if it already contains the content', function() {
-      var toInsert = 'blahzorz blammo';
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, toInsert, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert)
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal(toInsert, 'contents should be unchanged');
-          expect(result.inserted).to.equal(false, 'inserted should indicate that the file was not modified');
-        });
-    });
-
-    it('will insert into the file if it already contains the content if force option is passed', function() {
-      var toInsert = 'blahzorz blammo';
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, toInsert, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, { force: true })
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal(toInsert + toInsert, 'contents should be unchanged');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was not modified');
-        });
-    });
-
-    it('will insert into the file after a specified string if options.after is specified', function() {
-      var toInsert = 'blahzorz blammo';
-      var line1 = 'line1 is here';
-      var line2 = 'line2 here';
-      var line3 = 'line3';
-      var originalContent = [line1, line2, line3].join(EOL);
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, {after: line2 + EOL})
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal([line1, line2, toInsert, line3].join(EOL),
-                       'inserted contents should be inserted after the `after` value');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was modified');
-        });
-    });
-
-    it('will insert into the file after the first instance of options.after only', function() {
-      var toInsert = 'blahzorz blammo';
-      var line1 = 'line1 is here';
-      var line2 = 'line2 here';
-      var line3 = 'line3';
-      var originalContent = [line1, line2, line2, line3].join(EOL);
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, {after: line2 + EOL})
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal([line1, line2, toInsert, line2, line3].join(EOL),
-                       'inserted contents should be inserted after the `after` value');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was modified');
-        });
-    });
-
-    it('will insert into the file before a specified string if options.before is specified', function() {
-      var toInsert = 'blahzorz blammo';
-      var line1 = 'line1 is here';
-      var line2 = 'line2 here';
-      var line3 = 'line3';
-      var originalContent = [line1, line2, line3].join(EOL);
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, {before: line2 + EOL})
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal([line1, toInsert, line2, line3].join(EOL),
-                       'inserted contents should be inserted before the `before` value');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was modified');
-        });
-    });
-
-    it('will insert into the file before the first instance of options.before only', function() {
-      var toInsert = 'blahzorz blammo';
-      var line1 = 'line1 is here';
-      var line2 = 'line2 here';
-      var line3 = 'line3';
-      var originalContent = [line1, line2, line2, line3].join(EOL);
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, {before: line2 + EOL})
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal([line1, toInsert, line2, line2, line3].join(EOL),
-                       'inserted contents should be inserted after the `after` value');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(true, 'inserted should indicate that the file was modified');
-        });
-    });
-
-
-    it('it will make no change if options.after is not found in the original', function() {
-      var toInsert = 'blahzorz blammo';
-      var originalContent = 'the original content';
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, {after: 'not found' + EOL})
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal(originalContent, 'original content is unchanged');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(false, 'inserted should indicate that the file was not modified');
-        });
-    });
-
-    it('it will make no change if options.before is not found in the original', function() {
-      var toInsert = 'blahzorz blammo';
-      var originalContent = 'the original content';
-      var filePath = path.join(project.root, filename);
-
-      fs.writeFileSync(filePath, originalContent, { encoding: 'utf8' });
-
-      return blueprint.insertIntoFile(filename, toInsert, {before: 'not found' + EOL})
-        .then(function(result) {
-          var contents = fs.readFileSync(path.join(project.root, filename), { encoding: 'utf8' });
-
-          expect(contents).to.equal(originalContent, 'original content is unchanged');
-          expect(result.originalContents).to.equal(originalContent, 'returned object should contain original contents');
-          expect(result.inserted).to.equal(false, 'inserted should indicate that the file was not modified');
-        });
-    });
-  });
-
   describe('lookupBlueprint', function() {
     var blueprint;
     var ui;
